requirements.txt
cianparser
pandas
flask
jupyter
scikit-learn
matplotlib
black
dvc
flask-cors
boto3
python-dotenv
flask_httpauth 
locust